authenticatuion and autherization completed, save and apply for job completed

+++ src/controllers/company.controller.js
@@ -3,23 +3,15 @@ import Company from '../models/company.model.js';
 // Create a new company
 const createCompany = async (req, res) => {
   try {
-    const company = new Company(req.body);
+    const user = req.user.userId;
+    const company = new Company({ ...req.body, user });
     await company.save();
     res.status(201).json(company);
   } catch (error) {
-    res.status(500).json({ message: 'Server error', error });
+    res.status(500).json({ message: 'Server error', error: error.message });
   }
 };
 
-// Get all companies
-// const getCompanies = async (req, res) => {
-//   try {
-//     const companies = await Company.find();
-//     res.status(200).json(companies);
-//   } catch (error) {
-//     res.status(500).json({ message: 'Server error', error });
-//   }
-// };
 const getCompanies = async (req, res) => {
   const { name } = req.query;
 
@@ -32,7 +24,18 @@ const getCompanies = async (req, res) => {
     const companies = await Company.find(query);
     res.status(200).json(companies);
   } catch (error) {
-    res.status(500).json({ message: 'Server error', error });
+    res.status(500).json({ message: 'Server error', error: error.message });
+  }
+};
+
+
+const getCompaniesByAdmin = async (req, res) => {
+  try {
+    const user = req.user.userId;
+    const companies = await Company.find({ user });
+    res.status(200).json(companies);
+  } catch (error) {
+    res.status(500).json({ message: 'Server error', error: error.message });
   }
 };
 
@@ -46,34 +49,52 @@ const getCompany = async (req, res) => {
     }
     res.status(200).json(company);
   } catch (error) {
-    res.status(500).json({ message: 'Server error', error });
+    res.status(500).json({ message: 'Server error', error: error.message });
   }
 };
 
 // Update a company
 const updateCompany = async (req, res) => {
   try {
-    const company = await Company.findByIdAndUpdate(req.params.id, req.body, { new: true });
-    if (!company) {
+    const user = req.user.userId;
+    const companyToUpdate = await Company.findById(req.params.id);
+
+    if (!companyToUpdate) {
       return res.status(404).json({ message: 'Company not found' });
     }
-    res.status(200).json(company);
+
+    if (companyToUpdate.user.toString() !== user.toString()) {
+      return res.status(403).json({ message: 'Not authorized to delete this location' });
+    }
+
+    const updatedcompany = await Company.findByIdAndUpdate(req.params.id, req.body, { new: true });
+
+    res.status(200).json(updatedcompany);
   } catch (error) {
-    res.status(500).json({ message: 'Server error', error });
+    res.status(500).json({ message: 'Server error', error: error.message });
   }
 };
 
 // Delete a company
 const deleteCompany = async (req, res) => {
   try {
-    const company = await Company.findByIdAndDelete(req.params.id);
-    if (!company) {
+
+    const user = req.user.userId;
+    const companyToDelete = await Company.findById(req.params.id);
+
+    if (!companyToDelete) {
       return res.status(404).json({ message: 'Company not found' });
     }
+    if (companyToDelete.user.toString() !== user.toString()) {
+      return res.status(403).json({ message: 'Not authorized to delete this location' });
+    }
+
+    await Company.findByIdAndDelete(req.params.id);
+
     res.status(200).json({ message: 'Company deleted successfully' });
   } catch (error) {
-    res.status(500).json({ message: 'Server error', error });
+    res.status(500).json({ message: 'Server error', error: error.message });
   }
 };
 
-export { createCompany, getCompanies, getCompany, updateCompany, deleteCompany };
+export { createCompany, getCompanies, getCompany, updateCompany, deleteCompany, getCompaniesByAdmin };
